Move animations to animations-map
