+++ src/components/Chip/Chip.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import styled, { keyframes, css } from 'styled-components';
+import styled, { css } from 'styled-components';
 import Text from '../Text';
-import { palette } from '../variables';
+import { animations, palette } from '../variables';
 
 type Props = {
   text: string;
@@ -25,13 +25,6 @@ const Chip: React.FC<Props> = ({ text, isSelected, shouldShake, onToggle }) => {
     </StyledChip>
   );
 };
-
-const shakeAnimation = keyframes`
-  10%, 90% { transform: translate3d(-1px, 0, 0); }
-  20%, 80% { transform: translate3d(2px, 0, 0); }
-  30%, 50%, 70% { transform: translate3d(-4px, 0, 0); }
-  40%, 60% { transform: translate3d(4px, 0, 0); }
-`;
 
 const StyledChip = styled.button<{ isSelected: boolean; shouldShake: boolean }>`
   appearance: none;
@@ -61,8 +54,7 @@ const StyledChip = styled.button<{ isSelected: boolean; shouldShake: boolean }>`
   ${({ shouldShake }) =>
     shouldShake &&
     css`
-      animation: ${shakeAnimation} 0.82s cubic-bezier(0.36, 0.07, 0.19, 0.97)
-        both;
+      animation: ${animations.shake}
       backface-visibility: hidden;
       perspective: 1000px;
       transform: translate3d(0, 0, 0);
