+++ src/components/Spinner/Spinner.tsx
@@ -1,5 +1,5 @@
-import styled, { keyframes } from 'styled-components';
-import { palette } from '../variables';
+import styled from 'styled-components';
+import { animations, palette } from '../variables';
 
 type Variant = 'large' | 'medium' | 'small';
 
@@ -10,11 +10,6 @@ type Props = {
 export const Spinner: React.FC<Props> = ({ variant }) => {
   return <LoadingCircle id="loading" role="progressbar" variant={variant} />;
 };
-
-const spinAnimation = keyframes`
-from { transform: rotate(0deg); }
-to { transform: rotate(360deg); }
-`;
 
 const sizes: Record<Variant, number> = {
   large: 9.7,
@@ -35,15 +30,11 @@ const LoadingCircle = styled.div<{ variant: Variant }>`
       margin-bottom: ${diameter / 2}rem;
       `;
   }}
-  animation-duration: 1.5s;
-  animation-iteration-count: infinite;
-  animation-name: ${spinAnimation};
-  animation-timing-function: linear;
+  animation: ${animations.spin}
   border-bottom-color: ${palette.whiteOpacity};
   border-left-color: ${palette.whiteOpacity};
   border-radius: 50%;
   border-right-color: ${palette.whiteOpacity};
   margin-left: auto;
   margin-right: auto;
-  transition: all 0.5s ease-in;
 `;
