+++ src/components/variables.ts
@@ -1,3 +1,5 @@
+import { css, keyframes } from 'styled-components';
+
 export type Color = keyof typeof palette;
 
 export const palette = {
@@ -34,4 +36,41 @@ export const spacing = {
   layout_outer_spacing: '6vw',
   layout_spacing: '2rem',
   small_button_padding: '.5rem 2rem',
+};
+
+const growDown = keyframes`
+    0% {
+        transform: scaleY(0)
+    }
+    80% {
+        transform: scaleY(1.1)
+    }
+    100% {
+        transform: scaleY(1)
+    }
+`;
+
+const shake = keyframes`
+  10%, 90% { transform: translate3d(-1px, 0, 0); }
+  20%, 80% { transform: translate3d(2px, 0, 0); }
+  30%, 50%, 70% { transform: translate3d(-4px, 0, 0); }
+  40%, 60% { transform: translate3d(4px, 0, 0); }
+`;
+
+const spin = keyframes`
+from { transform: rotate(0deg); }
+to { transform: rotate(360deg); }
+`;
+
+export const animations = {
+  growDown: css`
+    ${growDown} 400ms ease-in-out forwards;
+  `,
+  shake: css`
+    ${shake} 0.82s cubic-bezier(0.36, 0.07, 0.19, 0.97)
+        both;
+  `,
+  spin: css`
+    ${spin} 1.5s linear infinite;
+  `,
 };
